REL-3609: add exptime param to ToO URL trigger
